Remove unused folder from csv format data and lap count from "raw"
Since images are now in the csv file itself, there is no need to
save them to a separate folder.

I had stopped saving images, but had not modified the init code that
created the directory for doing so in the first place. I have now done so.
Image directories are created for raw and ASL data only.

I also stopped creating an incrementing a lap counter for any data
format other than csv.