Add primitive line follower, adjust model input shape.
The line follower was is a poorly tuned exercise in PD control. Wobbly.

I have learned that the model.predict() method, which I had been using
to generate inputs from one frame of telemetry, is built for large batches
of data, and that simplying calling model(input) will give the predicted
values as desired (using the __call__ function)