Add dual output flag to config, autopilot
Keras is capable of building models that have one output each for
steering and throttle, and it can also output them as a tuple in a single
output. I am not sure there is much of a difference, but I've added code
to account for which way the model was trained.

It currently needs to be specified in config.py

There may be better ways to do it, but this works for now.