Automate model/scaler/telemetry processing
Previously, the model path, the scaler path, and the telemetry data
expected by the scaler had to be specified in config.py.

I have changed it so these are read from the model_history.csv file
produced by the Jupyter notebook that actually trains my models

The only command line argument required for autopilot is the index number
of the desired model.