Modify auto-mode driving setup.
Waiting for "is driving" to start inference led to a situation in which,
after pressing "drive", the client would then have to take the time to
load libcudnn and libcublas.

The client has now been reverted to before, which gets this out of the
way while waiting for the first command to start driving.